fix: moving directories, changing structure
