+++ src/layouts/landing-page/introduction/banner/index.tsx
