+++ src/layouts/landing-page/introduction/slider/Slider.module.css
@@ -47,51 +47,8 @@
   align-items: center;
 }
 
-.content {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin-right: 24px;
-  transition: all 300ms ease-in-out;
-  cursor: pointer;
-}
-
-.content:nth-child(2),
-.content:nth-child(3) {
-  opacity: 0.5;
-}
-
-.content:hover {
-  filter: brightness(1.15);
-}
-
-.content img {
-  width: 149px;
-  height: 80px;
-  object-fit: cover;
-  box-shadow: inset 0px 0px 1px #ffffff;
-  transition: all 300ms ease-in-out;
-}
-
-.content h2 {
-  font-weight: 700;
-  font-size: 12px;
-  color: #1d1a05;
-  margin-top: 5px;
-  text-transform: uppercase;
-  text-align: center;
-}
-
 .allImages {
   display: none;
-}
-
-@media (max-width: 1136px) {
-  .content img {
-    width: max-content;
-    height: 60px;
-  }
 }
 
 @media (max-width: 768px) {
@@ -131,19 +88,5 @@
 
   .allImages::-webkit-scrollbar {
     display: none;
-  }
-
-  .content {
-    justify-content: space-between;
-    width: 100%;
-  }
-
-  .content:nth-child(4) {
-    margin-right: 0;
-  }
-
-  .content img {
-    width: 140px;
-    height: 80px;
   }
 }
